requirements.txt
pyyaml==6.0.1
google-cloud-storage==2.10.0 